git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/branches/new-installer@53747 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ phase3/includes/specials/SpecialIpblocklist.php
@@ -162,7 +162,7 @@ class IPUnblockForm {
 	 * @return array array(message key, parameters) on failure, empty array on success
 	 */
 
-	static function doUnblock(&$id, &$ip, &$reason, &$range = null) {
+	static function doUnblock(&$id, &$ip, &$reason, &$range = null, $blocker=null) {
 		if ( $id ) {
 			$block = Block::newFromID( $id );
 			if ( !$block ) {
@@ -184,8 +184,7 @@ class IPUnblockForm {
 				if ( !$block ) {
 					return array('ipb_cant_unblock', htmlspecialchars($id));
 				}
-				if( $block->mRangeStart != $block->mRangeEnd
-						&& !strstr( $ip, "/" ) ) {
+				if( $block->mRangeStart != $block->mRangeEnd && !strstr( $ip, "/" ) ) {
 					/* If the specified IP is a single address, and the block is
 					 * a range block, don't unblock the range. */
 					 $range = $block->mAddress;
@@ -195,10 +194,21 @@ class IPUnblockForm {
 		}
 		// Yes, this is really necessary
 		$id = $block->mId;
+		
+		# If the name was hidden and the blocking user cannot hide
+		# names, then don't allow any block removals...
+		if( $blocker && $block->mHideName && !$blocker->isAllowed('hideuser') ) {
+			return array('ipb_cant_unblock', htmlspecialchars($id));
+		}
 
 		# Delete block
 		if ( !$block->delete() ) {
 			return array('ipb_cant_unblock', htmlspecialchars($id));
+		}
+		
+		# Unset _deleted fields as needed
+		if( $block->mHideName ) {
+			IPBlockForm::unsuppressUserName( $block->mAddress, $block->mUser );
 		}
 
 		# Make log entry
@@ -208,10 +218,9 @@ class IPUnblockForm {
 	}
 
 	function doSubmit() {
-		global $wgOut;
-		$retval = self::doUnblock($this->id, $this->ip, $this->reason, $range);
-		if(!empty($retval))
-		{
+		global $wgOut, $wgUser;
+		$retval = self::doUnblock($this->id, $this->ip, $this->reason, $range, $wgUser);
+		if( !empty($retval) ) {
 			$key = array_shift($retval);
 			$this->showForm(wfMsgReal($key, $retval));
 			return;
@@ -238,7 +247,7 @@ class IPUnblockForm {
 		$conds = array();
 		$matches = array();
 		// Is user allowed to see all the blocks?
-		if ( !$wgUser->isAllowed( 'suppress' ) )
+		if ( !$wgUser->isAllowed( 'hideuser' ) )
 			$conds['ipb_deleted'] = 0;
 		if ( $this->ip == '' ) {
 			// No extra conditions
@@ -306,7 +315,7 @@ class IPUnblockForm {
 	}
 
 	function searchForm() {
-		global $wgTitle, $wgScript, $wgRequest;
+		global $wgScript, $wgRequest, $wgLang;
 
 		$showhide = array( wfMsg( 'show' ), wfMsg( 'hide' ) );
 		$nondefaults = array();
@@ -330,11 +339,11 @@ class IPUnblockForm {
 		$links[] = wfMsgHtml( 'ipblocklist-sh-userblocks', $ubLink );
 		$links[] = wfMsgHtml( 'ipblocklist-sh-tempblocks', $tbLink );
 		$links[] = wfMsgHtml( 'ipblocklist-sh-addressblocks', $sipbLink );
-		$hl = implode( ' ' . wfMsg( 'pipe-separator' ) . ' ', $links );
+		$hl = $wgLang->pipeList( $links );
 
 		return
 			Xml::tags( 'form', array( 'action' => $wgScript ),
-				Xml::hidden( 'title', $wgTitle->getPrefixedDbKey() ) .
+				Xml::hidden( 'title', SpecialPage::getTitleFor( 'Ipblocklist' )->getPrefixedDbKey() ) .
 				Xml::openElement( 'fieldset' ) .
 				Xml::element( 'legend', null, wfMsg( 'ipblocklist-legend' ) ) .
 				Xml::inputLabel( wfMsg( 'ipblocklist-username' ), 'ip', 'ip', /* size */ false, $this->ip ) .
@@ -375,11 +384,10 @@ class IPUnblockForm {
 		if( is_null( $msg ) ) {
 			$msg = array();
 			$keys = array( 'infiniteblock', 'expiringblock', 'unblocklink', 'change-blocklink',
-				'anononlyblock', 'createaccountblock', 'noautoblockblock', 'emailblock', 'blocklist-nousertalk' );
+				'anononlyblock', 'createaccountblock', 'noautoblockblock', 'emailblock', 'blocklist-nousertalk', 'blocklistline' );
 			foreach( $keys as $key ) {
 				$msg[$key] = wfMsgHtml( $key );
 			}
-			$msg['blocklistline'] = wfMsg( 'blocklistline' );
 		}
 
 		# Prepare links to the blocker's user and talk pages
@@ -396,7 +404,7 @@ class IPUnblockForm {
 				. $sk->userToolLinks( $block->mUser, $block->mAddress, false, Linker::TOOL_LINKS_NOBLOCK );
 		}
 
-		$formattedTime = $wgLang->timeanddate( $block->mTimestamp, true );
+		$formattedTime = htmlspecialchars( $wgLang->timeanddate( $block->mTimestamp, true ) );
 
 		$properties = array();
 		$properties[] = Block::formatExpiry( $block->mExpiry );
@@ -418,7 +426,7 @@ class IPUnblockForm {
 			$properties[] = $msg['blocklist-nousertalk'];
 		}
 
-		$properties = implode( ', ', $properties );
+		$properties = $wgLang->commaList( $properties );
 
 		$line = wfMsgReplaceArgs( $msg['blocklistline'], array( $formattedTime, $blocker, $target, $properties ) );
 
@@ -434,7 +442,7 @@ class IPUnblockForm {
 
 			# Create changeblocklink for all blocks with exception of autoblocks
 			if( !$block->mAuto ) {
-				$changeblocklink = ' ' . wfMsg( 'pipe-separator' ) . ' ' .
+				$changeblocklink = wfMsgExt( 'pipe-separator', 'escapenoentities' ) .
 					$sk->link( SpecialPage::getTitleFor( 'Blockip', $block->mAddress ), 
 						$msg['change-blocklink'],
 						array(), array(), 'known' );
@@ -442,7 +450,7 @@ class IPUnblockForm {
 			$toolLinks = "($unblocklink$changeblocklink)";
 		}
 
-		$comment = $sk->commentBlock( $block->mReason );
+		$comment = $sk->commentBlock( htmlspecialchars($block->mReason) );
 
 		$s = "{$line} $comment";
 		if ( $block->mHideName )
